- products: view_count tiny->int 업데이트, cafe_id nullable, is_old 추가 - 네이버 메타 추가 - 옛 마켓관련 Disable - 메일, 이벤트 수정(주문 이벤트 추가) - 첨부파일 update 메소드 추가 - 중고 추가 - 네비 정보 추가

+++ database/migrations/2017_04_13_065958_create_products_table.php
@@ -21,7 +21,7 @@ class CreateProductsTable extends Migration
                 'road', 'mtb', 'fixed', 'hybrid', 'mini',
                 'bacc', 'racc', 'part'
             ]);
-            $table->tinyInteger('view_count')->default(0);
+            $table->unsignedInteger('view_count')->default(0);
             $table->string('ad_title');
             $table->enum('ad_status', ['준비중', '판매', '매진']);
             $table->string('ad_short_description')->nullable();
